Use sets to speed up illegal square checks
Users will rapidly hover over different walls.

Previously we were recalculating whether a square was legal or not
on each time a new square was hovered, which is a O(number of walls) operation.

Now we have two different sets which track the illegal indices for horizontal
and vertical walls respectively. We update these sets when new walls are placed.

Lookups are now O(1).